drivers: misc: adi: sram: Make SRAM compatible with vmap
The previous sram mmap driver used io_remap_pfn_range to map SRAM into
userspace. However this cannot be later used with an AF_XDP socket for
the umem buffer, because it creates a mapping with VM_IO that is not
backed by an instance of struct page.

This replaces the driver with one that uses a cooperating reserved
memory node to register SRAM as part of system memory and reserve it so
that it can later be mapped into userspace, used with AF_XDP, etc. but
will not be unexpectedly assigned to userspace otherwise.

The memory is also suitably mapped for use with POSIX IPC or any other
use which requires mapping it back into kernel space via get_user_pages,
and caching is enabled as usual for regular memory mapped into
userspace.